Created main, and added images
I created a main file that can be used to run our class. In main I cleaned up the code... a lot. Every line is commented and each chunk is explained. it is split up into methods that are called from each other, I probably shouldv've noted that with _func but its not a big deal. I also created two images, one for our title, it displays the words for the title and our logo twice, I kept having trouble using a frame so a single image works great there. The other image is for the button to get the weather. Also deleted a temporary image for testing.